requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
pydantic==2.13.3
joblib==1.5.3
scikit-learn==1.8.0
xgboost==3.2.0
pandas==2.3.3
numpy==2.4.4
pyyaml==6.0.2
mlflow==3.11.1
streamlit>=1.35.0
plotly>=5.22.0
httpx==0.28.1
pytest>=8.0.0
